refactor: extract UI/logic violations into hooks — round 2 (closes #1287)
- LyricsDisplay: extract SpeechSynthesis playback (lineIdx/wordIdx/playing + advance/cancel) → useLyricsPlayer.ts
- IsraelMapClient: extract feedback state + handleTap (store mutations, TTS, timers) → useIsraelMap.ts
- PWAInstallBanner: extract deferredPrompt state + beforeinstallprompt listener + install logic → usePWAInstallBanner.ts
- Fix bg-gradient-to-br → bg-linear-to-br (Tailwind v4 canonical classes) in both components
